[refactor] Car객체를 domain 패키지로 이동
