+++ src/main/java/study/step3/RacingGame.java
@@ -1,11 +1,10 @@
 package study.step3;
 
-import study.step3.dto.Car;
-import study.step3.dto.RacingInput;
+import study.step3.domain.Car;
 import study.step3.move.MoveStrategy;
-import study.step3.move.RandomlyMoveStrategy;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.List;
 
 public class RacingGame {
 
@@ -21,25 +20,6 @@ public class RacingGame {
 
     public List<Car> getCarList() {
         return carList;
-    }
-
-
-    public static void main(String[] args) {
-
-        InputView inputView = new InputView();
-        ResultView resultView = new ResultView();
-        RacingInput input = inputView.enter();
-
-        RacingGame racingGame = new RacingGame(input.getCarNum(), new RandomlyMoveStrategy());
-
-        for (int i = 0; i < input.getTryNum(); i++) {
-            racingGame.play();
-            List<Car> cars = racingGame.getCarList();
-            System.out.println("\n실행 결과");
-            resultView.printCars(cars);
-        }
-
-
     }
 
     public void play() {
